Support reusable comment generation.
- Add a helper to Google_Protobuf_SourceCodeInfo.Location to do the
  generation.
- Give ProvidesSourceCodeLocation a helper for getting the comments.
